# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/polizas": {
      "get": {
        "operationId": "PolizasController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "PolizasController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearPolizaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/polizas/{id}": {
      "get": {
        "operationId": "PolizasController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "put": {
        "operationId": "PolizasController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePolizaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "PolizasController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/polizas/{id}/amparos": {
      "get": {
        "operationId": "PolizasController_findAmparos",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/amparos": {
      "get": {
        "operationId": "AmparoPolizasController_findAll",
        "summary": "Listar todos los amparos",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "amparos"
        ]
      },
      "post": {
        "operationId": "AmparoPolizasController_create",
        "summary": "Crear múltiples amparos de pólizas",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "amparos"
        ]
      }
    },
    "/amparos/{id}": {
      "get": {
        "operationId": "AmparoPolizasController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "amparos"
        ]
      },
      "put": {
        "operationId": "AmparoPolizasController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAmparoPolizaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "amparos"
        ]
      },
      "delete": {
        "operationId": "AmparoPolizasController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "amparos"
        ]
      }
    }
  },
  "info": {
    "title": "Polizas API CRUD",
    "description": "API",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "API",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CrearPolizaDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePolizaDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateAmparoPolizaDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}